pyproject.toml
[tool.poetry]
name = "open-interpreter"
packages = [
    {include = "interpreter"},
]
version = "0.1.4"
description = "Let language models run code locally."
authors = ["Killian Lucas <killian@drinkwater.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.0"
rich = "^13.4.2"
tiktoken = "^0.4.0"
astor = "^0.8.1"
git-python = "^1.0.3"
tokentrim = "^0.1.9"
appdirs = "^1.4.4"
six = "^1.16.0"
python-dotenv = "^1.0.0"

# On non-windows systems, you can just `import readline`.
# On windows, `pyreadline3` replaces that, so you can also just `import readline`.
inquirer = "^3.1.3"
wget = "^3.2"
huggingface-hub = "^0.16.4"
litellm = "^0.1.720"
cassandra-driver = "^3.28.0"
[tool.poetry.dependencies.pyreadline3]
version = "^3.4.1"
markers = "sys_platform == 'win32'"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
interpreter = "interpreter:cli"
